+++ tests/lib/Integration/TestCase.php
@@ -87,6 +87,14 @@ abstract class TestCase extends BaseTestCase
     /**
      * @param Application $app
      */
+    protected function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('auth.providers.users.model', DummyApp\User::class);
+    }
+
+    /**
+     * @param Application $app
+     */
     protected function resolveApplicationExceptionHandler($app)
     {
         $app->singleton(ExceptionHandler::class, TestExceptionHandler::class);